Will Complete the First Exercis on Javascript.
This exercise was about  learning how to make repeating strings using
for loops. initially I got stuck writing the script, as I couldn't
figure how to write a for loop that would work with paramaters outside
of the loop itself. But I realized I had to nest the loop inside of a
if statement for it to do what I wanted it too. I also could not get the
loop initially to reach the rep statement, as it was outside of the
bounds of the the for statement, so it would repeatedly return an
empty string. I realized what was happening and properly placed the
rep expression inside of my for loop, and the addition stuck. Once I did
that I realized that the tests weren't treating each test as a separate
test, so I had to rewrite my expression so it would not do so.
Next time i will commit more so I can see the differences between code
when needed