Add Sum all to completed exercises
This one wasn't hard once I got the logic around
what was happening. What really helped was writing out
what was supposed to happen both in plain english and
in terms of the actual function using 1 - 4 as a test
case. Might help to get a pen and paper when I cant logically
figure out what something is supposed to do. was also good
for practicing nesting loops within if statements